% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw.R
\name{offence_trends}
\alias{offence_trends}
\title{OFFENCE TRENDS FINDER}
\usage{
offence_trends(crime_data, offence_description, postcodes)
}
\arguments{
\item{crime_data}{A data.table object with the following columns:
"date" (POSIXct), "suburb" (chr), "postcode" (chr), "offence_level_1" (chr),
"offence_level_2" (chr), "offence_level_3" (chr), "offence_count" (num).}

\item{offence_description}{A character string of <What are your expected inputs?>.}

\item{postcodes}{A two-element character vector. Each element is an SA postcode.}
}
\value{
A ggplot object showing the correlation in offence count between the two input postcodes.
}
\description{
\code{offence_trends} compares trends of frequency of chosen offence between two chosen passcodes by months
}
\examples{
offence_trends(crime_data,"THEFT AND RELATED OFFENCES", c["5041","5010"])
offence_trends(crime_data_table_as_above,"any_category_in_offence_level_1", a_vector_with_characters_elements_of_2_passcodes)
}
